provisioner/file-web-db/requirements.txt
Babel==2.9.0
click==7.1.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Table==0.5.0
Flask-WTF==0.14.3
IPy==1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
pytz==2021.1
PyYAML==5.4.1
Werkzeug==1.0.1
WTForms==2.3.3
